feat: Migrate React project to Next.js
This commit migrates the existing React (Vite) application to a standard, production-ready Next.js project.

The key changes include:
- Restructured the project to the standard Next.js format with `pages`, `components`, and `styles` directories.
- Replaced the Vite build system with Next.js scripts and dependencies.
- Moved the main UI from `App.tsx` to `pages/index.tsx`.
- Created `pages/_app.tsx` for global styles and `pages/_document.tsx` for custom document structure.
- Secured the Google Gemini API key by moving the API call to a backend API route at `pages/api/generate.ts`. The API key is now read from an environment variable on the server.
- Installed and configured Tailwind CSS to match the original project's styling.
- Removed obsolete files from the old Vite project structure.
