+++ pages/_document.tsx
@@ -1,0 +1,13 @@
+import { Html, Head, Main, NextScript } from 'next/document';
+
+export default function Document() {
+  return (
+    <Html lang="en">
+      <Head />
+      <body className="bg-[#000000] text-gray-100">
+        <Main />
+        <NextScript />
+      </body>
+    </Html>
+  );
+}
